Fixed problem 3

+++ src/main/java/com/nahidseidi/projecteuler/EulerProblem1.java
@@ -1,9 +1,9 @@
 package com.nahidseidi.projecteuler;
 
 public class EulerProblem1 implements IEulerProblem {
-   public int solve() {
-      int sum = 0;
-      for (int i = 0; i < 1000; i++) {
+   public long solve() {
+      long sum = 0;
+      for (long i = 0; i < 1000; i++) {
          if(i % 3 == 0 || i % 5 == 0) 
             sum += i; 
       }
