+++ src/main/java/com/nahidseidi/projecteuler/IEulerProblem.java
@@ -1,5 +1,5 @@
 package com.nahidseidi.projecteuler;
 
 public interface IEulerProblem {
-	public int solve();
+	public long solve();
 }
